+++ game-runner/src/v1_18_2/java/com/rettichlp/unicacityaddon/v1_18_2/VersionedPlayerListController.java
@@ -1,8 +1,8 @@
-package com.rettichlp.unicacityaddon.v1_20_1;
+package com.rettichlp.unicacityaddon.v1_18_2;
 
 import com.rettichlp.unicacityaddon.UnicacityAddon;
 import com.rettichlp.unicacityaddon.base.tab.TabPrefix;
-import com.rettichlp.unicacityaddon.controller.TabListController;
+import com.rettichlp.unicacityaddon.controller.PlayerListController;
 import net.labymod.api.Laby;
 import net.labymod.api.client.network.ClientPacketListener;
 import net.labymod.api.client.network.NetworkPlayerInfo;
@@ -22,11 +22,11 @@ import java.util.Map;
  * @author RettichLP
  */
 @Singleton
-@Implements(TabListController.class)
-public class VersionedTabListController extends TabListController {
+@Implements(PlayerListController.class)
+public class VersionedPlayerListController extends PlayerListController {
 
     @Inject
-    public VersionedTabListController() {
+    public VersionedPlayerListController() {
     }
 
     @Override
